chore(landing): moved landing component to admin/scenarios and changed to named export
