+++ next-tavla/pages/index.tsx
@@ -1,5 +1,5 @@
 import { Contrast } from '@entur/layout'
-import Landing from '../src/Landing'
+import Landing from '../src/Admin/scenarios/Landing'
 import classes from 'styles/landing.module.css'
 import { Header } from 'components/Header'
